+++ commands/misc/botinfo.js
@@ -1,0 +1,5 @@
+module.exports = {
+  name: "botinfo",
+  category: "misc",
+  run: async () => {},
+};