+++ commands/misc/invite.js
@@ -1,0 +1,5 @@
+module.exports = {
+  name: "invite",
+  category: "misc",
+  run: async () => {},
+};