+++ changelog.md
@@ -5,7 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.2] - 2021-05-03 [LATEST]
+## [0.0.3] - 2021-05-04 [LATEST]
+
+## Added
+
+- disconnect,filters,filter,loop,pause,resume, queuestats music commands,
+- botinfo,invite misc command
+
+## Updated
+
+- Distube Handler by adding Costom Filters
+
+### Removed
+
+- Unnessary Consloe logs from commands.
+
+### Fixed
+
+- client.prefix bug in ready event of bot
+
+## [0.0.2] - 2021-05-03
 
 ### Added